chore: controller 추가 - main은 프로그램의 실행만 담당할 수 있도록 controller를 추가하였다.
